Cache on-behalf-of tokens
After a long period of benchmarking and measurements it became clear
that a major contributor to the end-to-end response time was fetching
on-behalf tokens. In practice, it can be expected that fetching a fresh
on-behalf token 400-500ms, which for most requests is 2-30x the time it
takes to fetch data from blobs and build the response. For smaller
requests this can reduce the end-to-end latency from 5-600ms to 30ms.

The implementation uses go's built-in sync-map, and does not handle
token recovery, but for short-lived deployments it should work fine, and
demonstrate well effects of caching.