Invalidate tokens and retry on first failure
Check cached tokens by trying them, and re-trying once if the
token-dependent operation should fail. This pushes the responsibility
and expiration of tokens to whatever services the tokens were issued
for. This also means the paths where tokens are ok are faster, at the
cost of slower error paths. Since error paths involving third-party
services and tokens should quite rare, this should be a win.

The nature of the error is not checked, which means that perfectly valid
tokens will be discarded and re-requested if some external service
should fail for some other reason. I consider this perfectly acceptable
(at least for now), since the request already failed and is likely to
fail completely.