Automate the retry-on-token-rejection pattern
Manually implementing the nested error checks when retrying with fresh
tokens after a remote function fails is tedious and error prone, and
quite easily automated.

Please note that when implementing the approach by hand to test it, I
did mess it up once, and that was five minutes after coming up with the
approach (i.e. quite fresh in my memory). There is a significant tax on
immediate understanding, but it should eliminate a very common error and
make implementations more robust, and reduce the cost of dealing with
expired and rejected tokens.