Buffer the tasks and fragment channels
Use njobs-sized buffers for the tasks and fragment channels, which
improves throughput slightly by blocking less on channel read/write. The
fetch program relies on the channels for task distribution and rate
limiting, and not as a strict synchronisation semantics.

Some rough measurements indicates that just buffering channels reduces
the time blocked when waiting for new tasks from 2-60ms to 300µs-3ms.