Move shape-of-cube to base task description
In the process of designing a working /curtain endpoint, it became clear
that most, if not all, task descriptions will in some way or another
need the dimensions of the full survey. Until now it has been passed as
a part of the manifest, but it is inconvenient and leads to a lot of
extra parsing.

By promoting it to a part of the basic task description, the specific
ones can focus more on what sets them apart, and would have to write
less error prone parsing and serialization code.