Generalise plan/mkschedule
When I build the prototype for the /curtain endpoint it became quite
clear that adding new scheduling was a chore, and there was little
support for it. It turns out that the code was structurally identical to
the slice scheduling - really, I copy-pasted the implementation and
replaced the types.

This makes sense, really, because the thing in scheduling that depends
on the shape is the enriched parameters that goes along with the
fragment ID. For slices, this is just lineno & dimension, and for the
curtain it is a list of fragment-local dim0 and dim1 coordinate pairs.

From this, it seems that in order to add a new shape it should be
sufficient to write the shape-specific description, and not re-write
partitioning and task assembly.

In the context of just the slice endpoint this turns plan from being
somewhat straight forward to way over-engineered, and adds a significant
complexity tax. I do believe, however, that this will pay off in
maintenance and adding new endpoints.