Make scheduler in go a first-class type
Rather than the scheduler being an internal detail in the /slice
endpoint handler, make it a stand-alone first-class type. It reduces the
amount of code polluted by the awareness of cgo, and makes the scheduler
more formal and reusable.

With the exception of a few renames (because of the new context), this
is just a reorganisational change and does not affect functionality.