Build liboneseismic with fmt header-only
libfmt is an internal detail in oneseismic as it stands, and
liboneseismic is really not useful outside of linking with cgo. With
this in mind, the cgo does not have to link with libfmt. Linking against
a shared fmt is still certainly possible, as long as LDFLAGS is set in
accordance with the build environment.

A possible change for the future is to make fmt a purely internal
detail to oneseismic, and use FMT_HEADER_ONLY also in the cmake files.
However, it is overly restrictive and there are plenty environments
where users might want a shared fmt.