Perform scheduling (redis-write) from scheduler
Actually implement the scheduling (by writing objects to redis) from the
scheduler object, don't just generate the messages to be written.

This reduces the Slice.Get() function to "just" parsing the request
parameters and shuffling data into the next step in the pipeline, which
is a much better design than doing the scheduling inline.

Writing to redis from inside Slice.Get was always a temporary solution
to get some experience with a running system. All endpoints will
schedule its work the same way, so there is no reason to keep it as an
internal detail in Slice.