Move fetch/process from cgo to core/
This is largely a restructuring, moving the implementation from the cgo
sources to the core lib. This is in preparation for refactoring and some
re-design of the fetch/process communication.

The fmt::join() has been replaced with a hand-rolled join to avoid
depending directly on fmt in cgo. If libonseismic is built with
FMT_HEADER_ONLY=1 then cgo needs no other link than -loneseismic.