Rename process/slice request members
Rename the request members to input and output. They have been overall
awkward to name since they really do cover the same concept, but
different stages of the overall pipeline. By denoting them just with
the direction and leaving out the shape (in this case slice) the
relationship between them is clearer.

The shape is encoded in the type rather than the name.