Return fragment-ids as ;-delmited string
Instead of returning a vector of strings, the proc returns the
already-delimited list of ids. This means slightly less allocation
pressure and a nicer interface for the caller (cgo at the moment),
at the cost of being able to traverse this set without parsing/splitting
the string first.

With the add_fragment method in play, the fragments method no longer
needs to be virtual, and is implemented directly in proc. This means
derived procs needs to take care to actually populate the set of
fragments by calling add_fragment(), and can no longer just follow the
"template" of implementing all virtual calls.