Pre-allocate output buffers in proc/slice
Resize the slice output buffers rather than clearing or freshly
constructing them. This means the proc/slice can get its buffers filled
in parallel, since there is no longer a push back but rather an index'd
write. This is not yet leveraged from go.

There is a small security implication here: if an attack gets the
pointers and starts reading beyond the vector it might get residual data
from previous requests, but this does require an overflow to occur or
someone with access to the memory space of our process. Partial results
are (probably) not sensitive enough to warrant explicitly zeroing after
use. If that should be necessary the program should be rewritten to work
on fresh memory every time.

While this should result in less allocation pressure, the main
motivation is support for parallel writes to the buffers.