Add curtain support in core/
This patch adds features for reading curtains, defined as a a set of
full traces at arbitrary dim0/1 (also referred to as x/y) coordinates,
typically following a well path or a similar structure.

In terms of implementation, I've used a couple of tricks. The
fetch/process are made dumb workers and don't really know about columns
(fragments with the same dim0/1 ID, and changing 2/z), and just generate
whatever is requested. This generates more messages [1], but makes it
easier to adjust and scale work size. Work size should now be more in
line with /slice.

The proc/curtain relies on clear() and resize(), and allocates memory
early. While the current go code makes fresh instances every time, they
are designed to be longer-lived and reused. This comes at more state
that must be handled and a real tax on ease-of-read, but brings along
some speed improvements and, more importantly, support for parallel
add(). Pre-allocating buffers also makes the add() algorithm a lot
simpler.

The curtain_* messages demonstrate that the message design as it stands
is not well thought out, but this can be addressed later. The schema as
it is is reasonably simple and should give ample room for some
experience.

[1] if fetch was aware of columns, it would be sufficient to send x/y
    coordinates for a singe dim0/1 fragment, and let fetch generate IDs
    for the full column.