Access /curtain from python client
Add rough support for the /curtain endpoint from the python client. The
assembly code is awful and slow, and does show that the protocol is not
rich enough to build good and fast clients. Still, it seems to function
well, and should be good enough for some in-the-wild testing and
get us some good experience.