Add gvt.squeeze
I have recently experimented with adding new endpoints, and it turns out
the it is quite frequently a need to "collapse" into a new gvt with
lower dimenions, to easily compute the size and shapes of results. Doing
so is perfectly doable with the squeeze() on FS and CS, but in practice
you usually already have gvt available, and further results you need
depend on other gvt functions.