Use vmihailenco/msgpack for building result
Leverage a third-party vmihailenco/msgpack [1] library for building the
result header. The payload is already built inside oneseismic C++
library, but the go/api server is responsible for setting up the header
and "container" message structure.

There are a few options out there, but vmihailenco seemed the most
appropriate because it published the Encode() functions for data
structure prefixes, which is necessary when go never gets to see the
individual elements.

[1] vmihailenco/msgpack