Add basic_tuple::front, back
It is reasonably common to need to get the height/value of the "last"
dimension (typically trace size), regardless of how many dimensions
there are. The standard containers do come with an interface for this,
which we can leverage here.

The front() method is less interesting, but is included for symmetry.