Don't init unused schedule_makers
The compiler *should* be smart enough to realise it does not need to
construct the schedule_makers for the branches not taken, so efficiency
wise this change is probably a dud. Still, it makes sense to only
construct the schedule_maker that's about to be used, to not leak it
into unrelated scopes and contexts.