Add gvt.ndim
Add a user-available constant for the number-of-dimensions for a gvt
instance. The template ND parameter is only available inside the gvt
class, and while it is possible to retrieve it is a lot more clumsy than
the gvt just exposing it.

Most code should use gvt without caring too much about how many
dimensions there are, but it is sometimes useful to be able to look up
the number of dimensions for pre-allocations, avoid hard-coding
"last" dimensions, or avoid having to infer the same information from
other sources.