Add header() function to schedule_maker
To support efficient and elegant assembly of out-of-order results,
oneseismic is moving towards sending a small header first. This header
describes the shape, size, and keys of the index in sufficient detail
for clients to pre-allocate and traverse and "blit" directly into a
result data type. This depends on a bunch of shape and geometry
information, which should only available in the plan/C++ lib.

This commit includes the entry point and implementation for the curtain
and slice, but the header is removed before the data is being handed off
to go. This makes the change more focused and managable, since the go
code is unaware of the change.