Make parsing functions into methods; add .numpy()
Instead of a family of free assemble_* functions, pin them to classes.
The goal is to make the process class only aware of transport, process
status etc., and just delegates the parsing of the response itself to
some dependency.

Some minor renaming follows from this change, and the process handle is
bumped up significantly in importance. While this makes data slightly
less immediate, it does help solidify the idea that data from oneseismic
is not instant, and that processes can be started long before the data
is needed. This turns out to be very useful for real-world programs, at
the cost of having to call get() or numpy() to actually access the data
in python.