Add xarray() method for slice+curtain assembler
Implement the xarray() method that builds a labelled array. xarray seems
a good fit for oneseismic as a user, in particular when exploring data,
since it works with line numbers, labels, and coordinates natively,
meaning users won't have to manually query the oneseismic python
interface.

The xarray support is not complete, and labels and coordinates are
expected to change. Some of it is due to poor support in the message
_or_ lack of other endpoints with sufficient information to build good
labels. Still, this proves a good demo and a nice starting point, so the
blanks can be filled in later.