Flatten scheduler/tasks representation
Flatten the (cgo) struct tasks to use array-of-sizes and a large byte
array rather than an array of size+buffer objects. This simplifies the
copying code in go, and removes the need for a loop to free objects in
C++.

The size could have been woven into the byte array, and be parsed from
go, which would have saved another allocation. However, that would make
parsing more noisy and much more byte-oriented.