Compress (gzip) on /result
gzip the response if requested with /result/:pid?compress=gz. This is a
user per-request option as responses must be of a significant size [1]
for compression to make sense. If the caller is from inside the same
data centre as oneseismic, compression rarely makes sense.

There are some gin gzip middlewares available out there, but none really
fit. For oneseismic, we want to apply compression on request, whereas
the middlewares I found used file-type filters and/or size thresholds,
with poor or clumsy user defined filters. The implementation for gzip
(and for supporting more compression types) is sufficiently small for it
to make sense to just implement in util.

[1] At approx 20M the compression-time and time saved in transport seems
    to break-even between oneseismic and an azure VM.