Support compressed results in python client
Add support for compressed results. In terms of requests, it works by
adding a ?compression=gz query to the /result family of endpoints. The
requests library will automatically decode gzip if the Content-Encoding
header is set to gzip, so the oneseimic library is simply not aware.

This is an opt-in feature as compression is often a performance
degradation on fast networks.