Support wider range of tokens in blobfs-write
The blobfs-read logic was aware of different tokens effectively giving
different clients, and tried more interfaces before finally giving up.
The write-file assumed that the client was a an account-client and that
cwd was properly set, but this may not be the case if it is init'd from
a container-specific token.

Handling more cases is a (surprising to me) win in clarity since the
code paths of readable and writable handles can be unified, since it
*really* boils down to figuring out which client we have, and what
client we need.