Don't send failure-signal with explicit close
There is nothing, and should be nothing, that depends on a signal from
the failure channel for control flow. When the channel it closed, it
actually wakes up any blocking select, which as of now does nothing more
than print a log line:

| 2021/03/29 11:19:06 pid=be321e60-59c9-47f1-b8d5-bf0d7db5f49c, %!s(<nil>)

but could easily crash the application (or worse, not crash it) by
entring a failure handles when there is no failure.