Comment on why explicit channel close is necessary
Add a brief comment on why the explicit close(tiles) is necessary. There
is a real need for it since the explicit closing of channel *usually* is
wrong in go, except in this case where it controls some parts control
flow [1].

In this case, the open/closed status of tiles is *the* signal that the
loop should be considered completed, and task processing continue.

[1] for msg := range chan {}