Split scheduling into MakeQuery(); Schedule();
Rather than having schedule being a single call to the scheduler,
explicitly build a query plan, then dispatch it. Making the query is
performed first, and enables error checking in the plan builder, while
still allowing the scheduling itself (which means some I/O) to run off
in a separate goroutine. Additionally, the query plans can now be
traced, executed, tested, and simulated.

This is a small step towards properly formalising a query planner, with
input checking happening in C++. Go does not yet do anything more than
set InternalServerError in the status whenever something fails in C++ -
this is to come later.