Check if lineno exists from C++ only
The responsibility of checking was checked both in go and in C++, but
for different reasons. From the go perspective, the input is checked to
pass a properly validated request to C++, and to set the correct HTTP
status code. From the C++ perspective it's necessary to validate the
input anyway, if anything from a pure robustness standpoint.

With go being pushed towards a pure I/O system for oneseismic, the
input validation is moved to C++ and removed from go. The effect of this
patch is that requsts for lines not in the cube give a Internal Server
Error rather than a Not Found error, but this will change once the
communication between go and C++ gets to mature.