Send HTTP status code hint from C++ to MakeQuery
Since go never really inspects the user-provided input, it has no idea
on what error code to set, and must be told by C++. Add the status_code
field to the plan struct, which is either zero, or a proposed status
code.

With this patch, /slice set a 404 status code on bad line numbers.