Explicitly zero-initialize single plan; NRVO
Use a single, shared, named return value object when constructing a
plan, and zero-initialize it.

The explicit nullptr assignments are removed. This was not an obvious
change, since {} zeroing pointers is not as clear. Still, it makes it
harder to have uninitialised members and break the guarantee about err
and status_code being zero unless they are explicitly set.