Validate request parameters in C++
Perform the validation of input in C++, and only relay parameters and
report errors from go. This is another step in the move to making go a
pure I/O driver, and having all geometry and query logic in C++. It is
also a step in the direction towards a well defined query planner.

The C++ library has better tooling, support and know-how built into it
for dealing with geometry. Building a proper query planner module is
then natural to do in C++, which has the added benefit of completely
separating the responsibilities of the go and C++ code - go does I/O,
http, session management and concurrency, and C++ parses and validates
user parameters, handles geometry and builds the (internal) messages for
the other components in oneseismic.

Implementation wise this should come with few surprises, but it is worth
noting that the curtain schedule_maker builds the cartesian version
twice - once for the request itself, and once for the header. This is a
shortcoming in the design of the schedule maker, but fixing it is a
different change so it is postponed. This does highlight some problems,
and strengthens the argument for building a proper query planner with a
richer interface or custom language.