Add GraphQL support
This is a very exciting feature, introduced with a terrible
implementation.

Motivation
----------
So why this change? When playing around with various scenarios and
client implementations, it became quite clear that what oneseismic
really is deep down is a database engine. What you quickly want from
that is proper user driven queries, which you can build on top of the
REST URL-maps-to-resource scheme, but is quite cumbersome.

For ease of implementation, the approach was to have a very simple
interface for queries (since it has to be mapped to the concept of "the
resource"), and let clients handle mapping from whatever structure they
had into dimension/lineo, array-of-pairs etc. The more I work on this
the more I realise that you want some server support for easy-to-write
queries too.

What follows is that we want oneseismic to use a proper query language,
not some implicit language on a poorly documented schema of URLs. This
plays well with the recent developments where queries go through a
plan-and-optimise stage, before any data is actually being fetched.

All in all this introduces a new problem: building a query language.

Enter GraphQL
-------------
It's quite clear that oneseismic should not build its own query
language - it would have no support in popular languages, and building
and maintaining a good grammar and tooling is a massive task.
Fortunately, GraphQL seems a pretty good fit with a bunch of standard
tooling, and maps quite well onto the problem - it's a good third-party
solution for oneseismic.

The most important aspect is the introduction of a machine-readiable
schema. While REST APIs have stuff like OpenAPI, it was never really
introduced in oneseismic, and it is an "add-on". Of course there is
always a schema, but in oneseismic's case it was never formalised.

A quick demo
------------
This is the old (URL-based) workflow

GET oneseismic.equinor.com/query/0d235a7138104e00c421e63f5e3261bf2dc3254b/slice/0/152
response: {
    "authorization": "<key>"
    "location": "result/<pid>",
    "status": "result/<pid>/status"
}

And with this patch applied

GET oneseismic.equinor.com/graphql?{cube(id:"0d235a7138104e00c421e63f5e3261bf2dc3254b"){slice(kind:0,id:152){url, key}}}
response: {
    "data": {
        "cube": {
            "slice": {
                "url": "result/<pid>",
                "key": "<key>
            }
        }
    }
}

While both more verbose and deeply nested, it is parsed and expanded
using off-the-shelf libraries, and come with significant opportunities -
for example, all parameters are named and not url-position dependent,
and have support for enums:

query {
    cube(id: "0d235a7138104e00c421e63f5e3261bf2dc3254b") {
        slice(kind: inline, id:152) {
            url
            key
        }
    }
}

This is what the clients would then have to generate. This means clients
can also ask for a lot of other useful information, such as cube layout,
geometry, position in the same query as data is requested. On top of
that, GraphQL tools have support for exploring the schema, which means
good opportunities for evolving the data model. None of this is included
in this patch though, this only implements the model (and limitations)
in the /query based approach with GraphQL.

Implementation
--------------
The implementation is quite bad - it pretty much copies the
implementation of the /slice and /curtain endpoints into the GraphQL
handlers, and fixes compiler errors. This makes for a quick operational
/graphql endpoint without breaking the old /query endpoints, but an ugly
implementation with tons of warts. The graphql.go implementation should
be expected to be significantly rewritten with abstraction that fits the
new model.

From the looks of it, the C++ native tooling story for around graphql is
not stellar, but it seems alright in go. As a quick prototype go is a
good fit for this anyway, but at this point it is unclear if the Go <->
C++ communication protocol and data model should change, and at what
layer the proper business logic should live. Regardless, no change is
introduced with this patch, and the graphql resolvers simply produce the
same internal structures that the URL encoded schema did.

The way forward
---------------
I fully expect graphql to replace the REST model entirely, with a
notable exception: Response payloads are too slow, heavy, and
unpractical to be returned directly by the graphql, so the concept of a
oneseismic Promise type is introduced. This behaves like the promise
most (all?) programming languages, and is just the /result endpoint.
Clients will be aware of this, and use this information to fetch
(numerical) result of queries, although other query data may be returned
in the standard graphql manner.

The python libraries will have to be ported to call the graphql
endpoint, rather than the /query endpoint.

I expect the schema and input types to develop rapidly. I suspect that
the graphql implementations make high quality testing easier, because it
is a weak point in the test suite as of now.