Fetch slice, curtain with GraphQL in python
Port the slice and curtain calls to use the new graphql endpoint. This
is a pretty straight forward (with a tiny hack using multiple session
objects) port to the new backend, which makes the code somewhat messy
and the design somewhat poor.

I picked the gql library which has some promising features for its 3.x
series, but since it is not available on pypi without the --pre flag,
the graphql support is built on gql2.x. With the release of 3.x the
python module should be significantly overhauled, since some of the
fundamental assumptions have changed with the introduction of /graphql.