Automate auth-header-update in gql client
Extend the gql client using the same technique used for the
http_session/requests session, and implement execute so that it
automatically adds the oneseismic specific auth headers.

This removes the tedious self.session.headers and manual patching of the
auth header, at the cost of some magic in initialisation. This is a hack
that I expect will go away with a proper client redesign (based on the
graphql client and a different auth scheme).