Loosen type coupling in core/messages
The input and output messages from the scheduler shared a base type,
which is a stronger coupling than needs be. By splitting them and
duplicating some fields the relationship is a lot clearer, and the
messages much more flexible, which will be useful as the graphql schema
starts evolving.

This patch includes some renaming to clear up the concepts, but
functionally nothing is changed. From the scheduler, input types are
called *query, and output type *task.