Automate the generation of C++ pack() and unpack()
All leaf message types, both queries and tasks, are expected to come
with a pack() and unpack() pair, and with the powerful nlohmann/json
library all these implementations were identical (since nlohmann looks
for the to/from_json functions).

Managing these functions is a lot more work than it is worth, so auto
generate them instead with a mixin and some explicit instantiations. It
would be nice if the instantiations somehow could be automated, and not
be a manually managed set of instantiations, but it should be good
enough for now.