Send manifest as document, not embedded in string
Sending the document as a string for it to be parsed independently in
C++ was a nice hack which opened for the project to move on at the time,
but sending it as an embedded document proper is a better way forward.

This came with a (surprisingly) large number of changes, but they are
for the most part uninteresting.

The old message:

build {
    "pid": "<pid>",
    "token": "<token>",
    "manifest": "{ \"byteorder\": \"big\", \"dimensions\": [[...]] }",
    "function": "slice",
    ...,
}

becomes

build {
    "pid": "<pid>",
    "token": "<token>",
    "manifest": { byteorder: "big", "dimensions": [[...]] },
    "function": "slice",
    ...,
}

which makes parsing in C++ nicer, and generally fewer moving parts. The
overarching goal is to make golang not really inspect the manifest at
all, but only serve as an I/O layer, and have all parsing and validation
happen in C++.