Infer cube-shape from manifest in C++
Rather than parsing the manifest and embedding cube size in go, infer it
from the manifest in C++. This is a step in moving all validation into
C++ and making go a pure I/O layer.

The shape_cube field is gone from the basic_query, and subsequently
removed from message building in go.