Replace MakeTask with literals, handle args in C++
Build the query object (message.Task) with struct literals, and move the
unpacking of parameters to C++. This is a step to make go a pure I/O
layer, and do all model validation in C++.

Aside from syntactical changes, the big change is passing the unmodified
query args directly to C++ (as json), rather than converting the input
(list-of-pairs) to the working represesentation (pair-of-lists). The
conversion happens in the from_json(curtain_query) function.

Please note that the to_json serialization of curtain_query is removed.
It is unlikely that it will ever be interesting to actually render
queries into JSON from C++, and the query object has not yet set on a
canonical representation. To stop some easy-to-make errors, and since it
is unused anyway, the conversion is simply removed.