Split message.Task into Query, Task
Finally split the now properly distinct message types Query and Task,
which correspond to basic_query and basic_task in C++. This conceptual
difference was always there, but properly expressed in code. With the
cleanup in vocabulary comes some nice clarity, and the relationship
between the go and C++ message types are clearer.

With the new vocabulary and struct type, the "param" field has been
renamed to (the maybe more appropriate) args, which better maps onto
the GraphQL field(args: ...) syntax and semantics.