Split slice() into index, lineno aware versions
As of now, GraphQL does not support union input types, but there is a
proposal [1][2] for the next revision. It would be a welcome feature for
a system like oneseismic, because it could support referencing the same
structure (e.g. the slice) from representationally identical but
semantically different types. Consdier this example:

Schema:
    input Lineno { val: Int! }
    input Index  { val: Int! }
    slice(dim: Int!, idx: union Lineno | Index)

Query:
    slice(dim: 0, idx: Lineno { val: 1456 })
    slice(dim: 0, idx: Index  { val: 80 })

Both these queries would refer to the same slice. The best alternative
is really just separate queries for separate types, which in the
underlying implementation normalise their inputs before hooking into a
shared code path.

That is in essence what this patch does - it splits the slice() query
into a family of queries, with different functions taking different
reference systems. This allows users to query slices without having to
map to some inline, crossline or depth/time value, which makes
oneseismic a lot easier in many situations [3].

The validation and sanity checking is moved from build() into from_json,
so that build can expect a fully hydrated, validated, and normalised
query. Having this much model validation in parsing and nowhere else
points to a design flaw (should be dedicated functions for validation
that the parser can hook into), but that's a job for later.

The serializing code was only used for testing, and is removed as
to_json is not necessarily an inverse of from_json. It is not
functionally interesting to render a query back into json at this time.

[1] https://github.com/graphql/graphql-spec/blob/main/rfcs/InputUnion.md
[2] https://github.com/graphql/graphql-spec/issues/488
[3] e.g. when you have some UTM -> index map, but not UTM -> inline