Always get directory in localfs_from_args
The localfs_from_args helper is used both when setting up input (usually
a file name) and output (usually a directory), but when used like this:

oneseismic scan file.sgy

it would try to open file.sgy/file.sgy, which is wrong. localfs supplies
open() and really work on directories, so infer the directory from the
path even when it is a file.