Don't fetch manifest twice in linenumbers query
The cube setup will fetch the manifest, and it makes no sense for the
linenumbers resolver to fetch it again just to have it as []bytes to
parse it as json. In the same pass, distinguish different types of
errors and don't leak go panics into the graphql errors section. Any
error that happens in function is either temporal (e.g. blob connection
broke) or bad data in storage anyway, and little business for the end
user.

Converting from interface{} to [][]int is super awkward in go, so a
couple of helpers were introduced. Long-term this logic belongs in C++,
which has access to the fully-parsed document.