Redesign manifest
The manifest has for a very long time just been a copy of the scan
output, and used only to store the line numbers for a volume, but was
always intended to be the source of a richer set of parameters and
metadata. Since the rest of oneseismic has been too immature to leverage
a beefier manifest, it was never a priority, but we have reached the
point where it will be useful.

I would like to emphasize that I do not think this is a complete and
final "specification" for format-version 1, and that I expect a little
bit of a churn as we discover opportunities and problems.

Motivation
----------
This is a typical manifest so far:

    {
      "byteorder": "big",
      "format": 1,
      "samples": 5,
      "sampleinterval": 5000,
      "byteoffset-first-trace": 3600,
      "guid": "c26ff223460fb33230d2cb17bd2cdb95",
      "dimensions": [
        [
          5289,
          5290,
          5291,
        ],
          8361,
          8363,
          8365,
          8367,
        ],
        [
          0,
          5000,
          10000,
          15000,
          20000,
        ]
      ]
    }

Of all the data in here, only the guid and dimensions fields are
interesting, and it leaves a lot to be desired, for example:

* What are the available resolutions?
* What are the available partitions/fragmentations?
* Are there filters applied (e.g. compression)?
* Where is the survey?
* Are there other attributes (e.g. UTM coordinates) available?

The new design
--------------
The new design aims to capture all the things that oneseismic would want
to use to realise a query. Here is an example document:

    {
      "format-version": 1,
      "guid": "0d235a7138104e00c421e63f5e3261bf2dc3254b",
      "data": [
        {
          "file-extension": "f32",
          "filters": [],
          "shapes": [[64, 64, 64]],
          "prefix": "src",
          "resolution": "source"
        }
      ],
      "attributes": [
        {
          "type": "cdp",
          "layout":
          "tiled",
          "file-extension": "2f32",
          "labels": ["CDP X", "CDP Y"],
          "shapes": [[256, 512]],
          "prefix": "cdp"
        }
      ],
      "line-numbers": [
        [5289, 5290, 5291],
        [8361, 8363, 8365, 8367],
        [0, 5000, 10000, 15000, 20000]
      ],
      "line-labels": ["inline", "crossline", "depth"]
    }

The document includes a "format-version", which will allow gradual
migration and even concurrent format versions. This is a useful tool for
developing the format specification and gradually migrate data.

Rather than the data being implicit in "src/<partition>/<data.f32>",
this is now an explicit entry in an array of layouts. This is a
mechanism for oneseismic to decide what partitioning and resolution to
use, and makes it more robust w.r.t. paths. I expect that for most
volumes only one partitioning and one resolution would be available, but
this gives a useful tool for picking different layouts for different
needs.

The attributes key is new, and is intended for per-trace attributes such
as coordinates. The goal is for the entries to be self-descriptive and
support all kinds of attributes. Briefly, it opens up for queries like
this:

    {
      cube(id: ...) {
        sliceByIndex(dim: 0, index: 50, opts: { attributes: ["utm"] })
      }
    }

Which would include the UTM coordinates for the traces in the slice with
in the result package, should it be available.

Per-trace attributes can be stored and partitioned as a "squeezed"/flat
volume, and can be fetched and extracted using mostly the same
functionality as oneseismic already provides for volumes.

The "dimensions" key has been renamed to "line-numbers", and
"line-labels" is an explicit entry on the root. Line numbers are special
because all volumes need good line numbers, but they are quite local to
the file (unlike coordinates which usually are global). Line numbers can
also be stored as the set-of-numbers in either dimension, but this is
not a property of most attributes [0].

I considered making line-numbers just any other attribute, but nested
structures are less ergonomic to work with.

Conclusion
----------
All in all this makes the manifest a lot more powerful, with a natural
way of representing various layouts and attributes, and for data missing
or not yet supported by a specific version.

This patch only adds parsing (and upload) for the new manifest
prototype, but otherwise no new functionality is introduced - support
for the new manifest capabilities can be gradually added.

Absent from the patch is a migration strategy and script, but it is
really just a tiny program that fetches the (old) manifest, maps it onto
the new structure, and uploads it to the same address.

[0] for example, if the grid is even slightly rotated or squeezed, then
    coordinates can not be represented as the cartesian product of the
    unique values in all directions