Select fragment-shape from the manifest
The intention is for oneseismic to pick a shape for fragments that is
ideal w.r.t. caching, thickness for the query. Long-term it could accept
a hint, or even be smart about the available resolutions, but so far it
just picks the first available.

This patch is really about establishing the infrastructure for multiple
resolutions and multiple fragmentation shapes. In practice it will be
slightly more robust if some volume is uploaded at non-64x64x64.