Get prefix, ext dynamically in url building
Bits and pieces when building URLs for downloading has been hard coded,
as there has only been a few volumes available, no different resolution,
and only one data type, but it was always intended to be dynamic (i.e.
read from the manifest). This enables more data types and is preparation
for querying attributes, which come with different prefix and
extensions.