Move geometry implementation to headers
Implementing the geometry in a source file and explicitly instantiating
templates made sense at the time, as a compilation firewall for
dependencies and instantiation time. Now it is not so true anymore, and
having to explicitly instantiate all template classes and functions
makes gvt< 1 > impossible [0].

The main benefit of the compilation firewall was for language bindings,
but that may never actually materialise. Currently, all use is internal,
that is, no header includes geometry.hpp, which drastically reduces the
value of implementing the templates in a source file.

    $ grep -rn "geometry.hpp" api/ core/
    tests/process.cpp:3:#include <oneseismic/geometry.hpp>
    tests/process.cpp.orig:3:#include <oneseismic/geometry.hpp>
    tests/geometry.cpp:3:#include <oneseismic/geometry.hpp>
    include/oneseismic/geometry.impl.hpp:13:#include <oneseismic/geometry.hpp>
    src/plan.cpp:10:#include <oneseismic/geometry.hpp>
    src/process.cpp:7:#include <oneseismic/geometry.hpp>

When it comes to compilation time, it is not so bad either. Since
geometry.cpp instantiated all templates, it is the worst case:

    $ time make src/geometry.o"
    Building CXX object CMakeFiles/oneseismic.dir/src/geometry.cpp.o
    real    0m2.092s
    user    0m1.958s
    sys     0m0.116s

All in all it is worth moving the implementation to the headers for
ease-of-build and for supporting gvt<1>, with higher dimensions "just
working".

[0] by refactoring out a basic_gvt with template specialisations for 1
    (without squeeze) and ND > 1 (with squeeze) it would work